config/requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
distutils-pytest==0.2.1
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
idna==3.10
iniconfig==2.0.0
logging==0.4.9.6
magic-filter==1.0.12
mechanize==0.4.10
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
PySocks==1.7.1
pytest==8.3.4
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
yarl==1.18.3
